requirements.txt
# local package
-e .

# external requirements
click~=7.1.2
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

pandas~=1.1.2
scikit-learn~=0.23.2
numpy~=1.19.1
bs4~=4.9.3
beautifulsoup4~=4.9.3
setuptools~=49.6.0
nltk~=3.5